examples/chat/pyproject.toml
[tool.poetry]
name = "chat"
version = "0.1.0"
description = ""
authors = ["Alasdair <alnessy@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
phoenix = { path = "../../" }
aiohttp = "^3.6.2"
attrs = "^19.3.0"
multipledispatch = "^0.6.0"
cattrs = "^1.0.0"
pyrsistent = "^0.16.0"
pytz = "^2020.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^19.10b0"

[tool.poetry.scripts]
server = "chat.__main__:server"
client = "chat.__main__:client"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
